Rename SayHello to TestProcess and simplify process
