--- processes/__init__.py
@@ -1,5 +1,0 @@
-from .wps_say_hello import SayHello
-
-processes = [
-    SayHello(),
-]
