+++ tests/processes/__init__.py
@@ -1,0 +1,5 @@
+from .wps_test_process import TestProcess
+
+processes = [
+    TestProcess(),
+]
